requirements.txt
asgiref==3.5.0
Django==4.0.3
djangorestframework==3.13.1
Pillow==9.0.1
pytz==2022.1
searchform==0.1.1
sqlparse==0.4.2
view==0.1
